Basic Setup
I have set up the basic look of the api. There will be main route files, which will include all the routes. All of the functions/methods will be contained in their own controller file. I have yet to create a models folder. However, a models folder will include all of the schemas for the different types of data.