Added Dynamic Sorting
Not sure if I can call it dynamic sorting. But, I added a filter option. On the dashboard, I want to show the the categories in order from created on the categories page. Then, on the sub categories page to find things easier, I want the categories to be sorted by alphabetization. So, I created a filter option, so both of those now work.