Product Route Changes
I took a look at the routes for products and realized some issues with them. Firstly, a few of the routes would return all the products no matter if they are set to not be active. This is an issue when opening the store and customers can buy products that aren't meant to be sold yet. Further, I have fixed an issue on the seasonal products controller that I believe may have rendered that route useless. I have also added a check to only find products that are active on the two main product and seasonal product routes. I also created a new route that will return all products no matter the status of being active or the season. This route requires admin authentication, because it will show ALL products inlcuding those that you may not want the customer to see. Lastly, I have added some comments above two of the controllers just outlining some basic information about each controller.