Updated Products Route
I have updated the products route to be more flexible. It now is able to accept sorting by ascending or descending. These can be applied to the created date or updated date. This is extremely useful. It allows you to be able to have a section for new arrivals, updated/upgraded products, old products, etc.