Delete Images + Main Image
Now, when you delete a product, it also deletes the images instead of just leaving them. This will help to save resources. I also added a new field for shelf life and for a main image. It took 2 hours, but I finally got the main image section working. You can upload a single image as the main image and this will be displayed on the product card.