Minor Changes
I came across a bug where for some reason the products per page was defaulting to 4. I found out that what was being sent was sent as a string and not number/integer. So, I just added a conversion to make sure that it is indeed passing an integer and not string.