Change to Individual Product
With changes to the front end, I have modified the backend route to get an individual product. Foreseeing changes, I already added a route prevously specifically for admin to get information of a product that was going to be updated. This way, they could get all the necessary information. However, the user shouldn't be allowed to know how many of an item have been sold, what the company pays for an item, or the suggested quantity in stock. These fields that may be more business related will be sorted out.