Ratings System
Started the rating system. There is a route that will go to a controller method for updating or adding a review. It will grab information relating to the user, product, as well as req.body. From here, it will check to see if the user has already authored a review of that product. If they haven't, it will create a new rating, review, or both on that product for that user. If they have already created a review or rating, it will simply update their old one.